% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_gps.R
\name{interp_cleanup}
\alias{interp_cleanup}
\title{Clean up after interpolation}
\usage{
interp_cleanup()
}
\description{
Removes temporary environment (.exif) containing data that was used
during interpolation.
}
\details{
As closing R session will also delete that environment,
this is not mandatory.
}
