% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_gps.R
\name{interp_josm}
\alias{interp_josm}
\title{Interpolation and JOSM update}
\usage{
interp_josm(path = ".")
}
\arguments{
\item{path}{(optional) path where input and output are located.
This is not necessary if the working directory is already set to
that location.}
}
\value{
Nothing. Only \code{.exif$base} and \code{approx.osm} are updated.
}
\description{
Reads exact positions, linearly interpolates non-exact positions,
updates osm file and displays it in JOSM.
}
\details{
Reads \code{approx.osm} and detects exact locations (moved manually in JOSM).
Then interpotales linearly other points. Updates \code{approx.osm} and
opens it in JOSM (http://localhost:8111 is used to open the file in JOSM,
which must be opened beforehand).
}
\examples{
# interp_josm()
}
