% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_coord_csv}
\alias{write_coord_csv}
\title{Write coordinates and direction in a csv file.}
\usage{
write_coord_csv(path = ".", file = "interp_gps.csv", direction = TRUE, ...)
}
\arguments{
\item{path}{(optional) path where input and output are located.
This is not necessary if the working directory is already set to
that location.}

\item{file}{the name of the csv file (default : "interp_gps.csv")}

\item{direction}{should direction to next photo be calculated and included
in the file?}

\item{...}{other arguments passed to \code{write.csv}}
}
\description{
Write computed longitudes, latitudes (and directions) in a csv file.
Images are not modified.
}
