% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_exiftool}
\alias{write_exiftool}
\title{Writing exif tags with exiftool}
\usage{
write_exiftool(path = ".", direction = TRUE)
}
\arguments{
\item{path}{(optional) path where input and output are located.
This is not necessary if the working directory is already set to
that location.}

\item{direction}{should direction to next photo be calculated and included
in the file?}
}
\description{
Runs \href{https://sno.phy.queensu.ca/~phil/exiftool/}{exiftool}
commands to write computed longitudes, latitudes (and directions) in image
files.
}
\details{
In order to write image files, \strong{exiftool} must be installed.

Files are not overwritten : a copy of the images including new exif tags
are written in the \code{output} directory.
}
