% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_exiftool_bat}
\alias{write_exiftool_bat}
\title{Create a batch file for exiftool}
\usage{
write_exiftool_bat(path = ".", file = "write_exiftool.bat",
  direction = TRUE)
}
\arguments{
\item{path}{(optional) path where input and output are located.
This is not necessary if the working directory is already set to
that location.}

\item{file}{the name of the batch file (default : "write_exiftool.bat")}

\item{direction}{should direction to next photo be calculated and included
in the file?}
}
\description{
Create a batch file that contains \href{https://sno.phy.queensu.ca/~phil/exiftool/}{exiftool}
commands to write computed longitudes, latitudes (and directions) in image
files.
}
\details{
In order to execute the batch file created by that function, \strong{exiftool}
must be installed.

Files are not overwritten : a copy of the images including new exif tags
are written in the \code{output} directory.
}
